Extract player control block into ui::control_box
Move the title row, seekbar, transport strip, status line, glyph helpers,
and the eq/streaming dropdowns out of the player view into a shared
control_box module, rendered through render_at() over the view's legacy
rows so output stays byte-identical. Groundwork for the bottom-docked
control box; also relieves the player.rs size-ratchet pin (1698 -> 788).
